split esplora files
